split MetadtaNodeVisitor to standalone visitors
