+++ packages/NodeTypeResolver/src/NodeVisitor/NamespaceNodeVisitor.php
@@ -1,15 +1,15 @@
 <?php declare(strict_types=1);
 
-namespace Rector\NodeTypeResolver\Metadata;
+namespace Rector\NodeTypeResolver\NodeVisitor;
 
 use PhpParser\Node;
 use PhpParser\Node\Stmt\Namespace_;
 use PhpParser\Node\Stmt\Use_;
-use Rector\NodeTypeResolver\Contract\Metadata\NodeDecoratorInterface;
+use PhpParser\NodeVisitorAbstract;
 use Rector\NodeTypeResolver\Node\Attribute;
 use Rector\Utils\BetterNodeFinder;
 
-final class NamespaceNodeDecorator implements NodeDecoratorInterface
+final class NamespaceNodeVisitor extends NodeVisitorAbstract
 {
     /**
      * @var string|null
@@ -36,14 +36,20 @@ final class NamespaceNodeDecorator implements NodeDecoratorInterface
         $this->betterNodeFinder = $betterNodeFinder;
     }
 
-    public function reset(): void
+    /**
+     * @param Node[] $nodes
+     * @return Node[]|null
+     */
+    public function beforeTraverse(array $nodes): ?array
     {
         $this->namespaceName = null;
         $this->namespaceNode = null;
         $this->useNodes = [];
+
+        return null;
     }
 
-    public function decorateNode(Node $node): void
+    public function enterNode(Node $node): ?Node
     {
         if ($node instanceof Namespace_) {
             $this->namespaceName = $node->name ? $node->name->toString() : null;
@@ -54,5 +60,7 @@ final class NamespaceNodeDecorator implements NodeDecoratorInterface
         $node->setAttribute(Attribute::NAMESPACE_NAME, $this->namespaceName);
         $node->setAttribute(Attribute::NAMESPACE_NODE, $this->namespaceNode);
         $node->setAttribute(Attribute::USE_NODES, $this->useNodes);
+
+        return $node;
     }
 }
